# docs/swagger.yaml
definitions:
  handler.request:
    properties:
      category:
        type: string
      count:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: This API Handle MELI Products.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /products:
    delete:
      consumes:
      - application/json
      description: Delete product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: Update products name
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update Name
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Update products
      tags:
      - Products
swagger: "2.0"